project/requirements.txt
Flask
cryptography
pycryptodome
dnslib